update post but now delete buttons dont work

+++ src/pages/ForumPage.jsx
@@ -27,7 +27,7 @@ function ForumPage() {
         setContent('');
         fetchPosts();
         } catch (error) {
-        console.log(error);
+            console.log(error);
         }
     }
 
@@ -39,6 +39,11 @@ function ForumPage() {
         setContent(e.target.value);
     }
 
+
+
+
+
+
     return (
         <div>
         <h1>Forum</h1>
@@ -46,7 +51,7 @@ function ForumPage() {
         <div>
             <form method="POST" action="/createpost" onSubmit={handleCreatePost}>
             <div>
-                <label htmlFor="title">Title:</label>
+                <label>Title:</label>
                 <input
                 type="text"
                 id="title"
@@ -55,7 +60,7 @@ function ForumPage() {
                 />
             </div>
             <div>
-                <label htmlFor="content">Content:</label>
+                <label>Content:</label>
                 <input
                 id="content"
                 value={content}
@@ -67,17 +72,21 @@ function ForumPage() {
         </div>
         <div>
             {posts.map((post) => (
-            <Post key={post._id} post={post} />
+            <Post key={post._id} post={post} setPosts={setPosts}/>
             ))}
         </div>
         </div>
     );
 }
 
-function Post({ post }) {
+function Post({ post, setPosts }) {
     const [comments, setComments] = useState([]);
     const [commentContent, setCommentContent] = useState('');
     const {token} = useContext(SessionContext);
+
+    const [isEditing, setIsEditing] = useState(false);
+    const [updatedTitle, setUpdatedTitle] = useState(post.title);
+    const [updatedContent, setUpdatedContent] = useState(post.content);
 
     useEffect(() => {
         fetchComments();
@@ -92,15 +101,13 @@ function Post({ post }) {
         e.preventDefault();
         try {
             await axios.post(`http://localhost:5005/forum/posts/${post._id}/createcomment`, 
-                {
-                    content: commentContent,
-                },
-                {headers: { Authorization: `holder ${token}`  } }
-                );
+                { content: commentContent }, 
+                { headers: { Authorization: `holder ${token}`  } }
+            );
             setCommentContent('');
             fetchComments();
         } catch (error) {
-        console.log(error);
+            console.log(error);
         }
     }
 
@@ -108,42 +115,113 @@ function Post({ post }) {
         setCommentContent(e.target.value);
     }
 
+    
+// delete post
     async function handleDeletePost() {
+        if (post.author._id !== token.userId) {
+            console.log("You're not authorized to delete this post");
+            return;
+        }
+
         try {
             await axios.delete(`http://localhost:5005/forum/posts/${post._id}`, {
                 headers: { Authorization: `holder ${token}` },
             });
+
+            setPosts((prevPosts) => prevPosts.filter((p) => p._id !== post._id));  
         } catch (error) {
             console.log(error);
         }
     }
 
-    async function handleDeleteComment() {
+//update post
+    async function handleUpdatePost(e) {
+        e.preventDefault();
         try {
-            await axios.delete(`http://localhost:5005/forum//posts/${post._id}/comments/${comment._id}`, {
-                headers: { Authorization: `holder ${token}` },
-            });
+            await axios.patch(`http://localhost:5005/forum/posts/${post._id}`, 
+                { title: updatedTitle, content: updatedContent },
+                { headers: { Authorization: `holder ${token}` } }
+            );
+            setIsEditing(false);
         } catch (error) {
             console.log(error);
         }
+    }
+
+
+
+    async function handleDeleteComment(commentId) {
+        const commentToDelete = comments.find((c) => c._id === commentId);
+        if (commentToDelete.author._id !== token.userId) {
+            console.log("You're not authorized to delete this comment");
+            return;
+        }
+
+
+        try {
+            await axios.delete(`http://localhost:5005/forum/posts/${post._id}/comments/${commentId}`, {
+                headers: { Authorization: `holder ${token}` },
+            });
+            setComments((prevComments) => prevComments.filter((comment) => comment._id !== commentId));
+        
+        } catch (error) {
+            console.log(error);
+        }
+    }
+
+    function handleUpdatedTitleChange(e) {
+        setUpdatedTitle(e.target.value);
+    }
+
+    function handleUpdatedContentChange(e) {
+        setUpdatedContent(e.target.value);
     }
 
     return (
         <div>
-        <h2>{post.title}</h2>
-        <p>{post.content}</p>
-        <p>{post.author.username}</p>
+        {isEditing ?  (
+            <form onSubmit={handleUpdatePost}>
+                <div>
+                    <label>Title:</label>
+                    <input
+                        id="title"
+                        value={updatedTitle}
+                        onChange={handleUpdatedTitleChange}
+                    ></input>
+                </div>
+                <div>
+                    <label>Content:</label>
+                    <textarea
+                        id="content"
+                        value={updatedContent}
+                        onChange={handleUpdatedContentChange}
+                    ></textarea>
+                </div>
+                <button type="submit">Update Post</button>
+                <button onClick={handleDeletePost}>Delete Post</button>
+            </form>
+    ) : (
+        <>
 
+
+
+        <h2>Post title: {post.title}</h2>
+        <p>Post content: {post.content}</p>
+        <p>Post author: {post.author.username}</p>
+
+        <button onClick={() => setIsEditing(true)}>Edit Post</button>
         <button onClick={handleDeletePost}>Delete Post</button>
+        </>
+    )}
 
         <div>
             <form method="POST" action="/posts/:postId/createcomment" onSubmit={handleCreateComment}>
                 <div>
-                    <label htmlFor="content">Comment:</label>
+                    <label>Comment:</label>
                     <input
-                    id="content"
-                    value={commentContent}
-                    onChange={handleCommentContentChange}
+                        id="content"
+                        value={commentContent}
+                        onChange={handleCommentContentChange}
                     ></input>
                 </div>
             <button type="submit">Create Comment</button>
@@ -153,9 +231,9 @@ function Post({ post }) {
         <div>
             {comments.map((comment) => (
                 <div key={comment._id}>
-                    <p>Comment: {comment.content}</p>
-                    <p>By: {comment.author.username}</p>
-                    <button onClick={handleDeleteComment}>Delete Comment</button>
+                    <p>Comment content: {comment.content}</p>
+                    <p>Comment By: {comment.author.username}</p>
+                    <button onClick={() => handleDeleteComment(comment._id)}>Delete Comment</button>
                 </div>
             ))}
         </div>
